Cannot able to delete the field in Exploratory testing (#52)
* For an issue #47

While deleting the field from the dialog call "DeleteField" method it takes module name and field name as input. So fetch those values from the combo box and from table. The field header should not be reordered :) so enforced "false" in field table.

Added null check on the column values and removed in JSON file